+++ doc/changelog.rst
@@ -1,6 +1,6 @@
 Changelog
 =========
-
+* :support:`4` Add encryption and decryption for caesar cipher
 * :support:`8` Add contributing guidelines.
 * :support:`3` Add Travis build status, code coverage, and Read The
   Docs badges.